+++ git-pack/CHANGELOG.md
@@ -1,3 +1,6 @@
 ### 0.9.0 (2021-08-??)
 
-- TBD
+- **renames / moves**
+   - `find::Find`  and `find::FindExt` only in `Find` and `FindExt` (not in `find` anymore)
+   - `data::output::count::Count` -> `data::output::Count`
+   - `data::output::entry::Entry` -> `data::output::Entry`